Timestamp: fix timezone preservation test, and depend on new MooseX::Timestamp
There was a bug in MooseX::TimestampTZ coercion rule, it would lose the
input timezone and convert to local time.  I didn't notice, because the
test used a timezone which happened to match my machine's timezone rules.
Use a bizarre timezone for the test, and depend on the version of
MooseX::Timestamp which fixes the bug (just released).