wr72818 - Error method to return the ErrorDetails as a (name => value) list
The 'details' method simply returns all of the values that were passed back
from the SRS; however there is little consistency to these.  The intent of
the table of error fields was to be able to annotate the values that came
back with what they mean.  Add a function which implements this, as well as
an extra transform for errors which in turn a 'value', 'fieldName', ...
list.