wr77087: loosen up type checking a little bit
Because the definition of fields such as 'status' is controlled by the
backend, but XML::SRS is (currently) only used on the client, there's
potential for the SRS to introduce new allowed values, which XML::SRS
believes are invalid. This would result in client-side validation failures.
To avoid this, we make the validation of a couple of fields less restrictive.
(There are possibly others than need the same treatment).

When/if XML::SRS is used by the backend, we'll then have one place where
the validation is defined, and so can become stricter.