wr77086: remove Crypt::Password, as it is not platform independent.
Crypt Password uses "Modular Crypt Format", which is specific to platforms
(at least is different on linux and bsd). This makes it very difficult for
clients to construct encrypted passwords that the SRS can make use of. It
also makes this module difficult to install on non-linux platforms. However,
since the SRS supports plain-text passwords (which it then converts itself),
we can just remove Crypt::Password altogether.